Weight transactions (#48)
* Weight transaction models

* Fixed date migrations

* Weight services

* API endpoints for Weight Transactions

* Weight FE scaffolding

* Weigh transaction unit tests

* Weight state and unit test fixes

* Weight transaction form

The weight transaction form is a component that is focussed
on capturing the information relating to a weight transaction.
It will be hosted on both new and edit pages. It is a pure
component only concerned with rendering logic.

The pages, edit and new are located at routes:
- new: `/weight/new`
- edit: `/weight/:id/edit`
They are responsible for managing the data fetch requests
and async operations.

+++ livestock-tracker.logic/Services/Feed/FeedingTransactionCrudService.cs
@@ -4,6 +4,7 @@ using LivestockTracker.Abstractions.Models.Feed;
 using LivestockTracker.Abstractions.Services.Feed;
 using LivestockTracker.Database;
 using LivestockTracker.Database.Models.Feed;
+using LivestockTracker.Logic.Paging;
 using LivestockTracker.Models.Paging;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
@@ -160,7 +161,7 @@ namespace LivestockTracker.Logic.Services.Feed
         /// <param name="key">The unique identifier value for the feeding transaction.</param>
         /// <param name="cancellationToken">A token that can be used to signal operation cancellation.</param>
         /// <returns>The matching feeding transaction if found.</returns>
-        public virtual async Task<IFeedingTransaction?> GetOneAsync(int key, CancellationToken cancellationToken)
+        public virtual async Task<IFeedingTransaction?> GetOneAsync(long key, CancellationToken cancellationToken)
         {
             Logger.LogInformation($"Finding a feeding transaction that matches ID {key}...");
             var feedingTransaction = await LivestockContext.FeedingTransactions
@@ -177,7 +178,7 @@ namespace LivestockTracker.Logic.Services.Feed
         /// <param name="key">The unique identifier value for the feeding transaction.</param>
         /// <param name="cancellationToken">A token that can be used to signal operation cancellation.</param>
         /// <returns>The ID of the removed feeding transaction.</returns>
-        public virtual async Task<int> RemoveAsync(int key, CancellationToken cancellationToken)
+        public virtual async Task<long> RemoveAsync(long key, CancellationToken cancellationToken)
         {
             Logger.LogInformation($"Deleting feeding transaction with ID {key}...");
             var feedingTransaction = await LivestockContext.FeedingTransactions
